refactor: create wss class for creating websocket server and handlibg client requests
